# internal/handler/static/openapi.yaml
openapi: 3.0.3
info:
  title: Simple Host API
  description: >-
    Static website hosting with a light per-site backend. Deploy a site with one
    API call — inline JSON files (LLM-friendly) or a tar.gz/zip archive — and each
    site gets shared JSON state (with atomic ops), append-only collections. Sign-in is a 6-digit email code or Google that
    returns an API key, sent as the
    `X-API-Key` header. Hosted-page sign-in uses the same account but a
    site-scoped session cookie on the site's own address — never an API key.
    Every site lives at its own address, `https://<site>.<handle>.simple-host.app/`
    (its own browser origin); `https://<handle>.simple-host.app/` lists the
    person's public sites. For a brand-new account a site may briefly live at
    `https://<handle>.simple-host.app/<site>/` until its certificate is issued:
    use the `site_url` the API returns, never a composed one.
    Old `<handle>.simple-host.app/<site>/` links redirect to the site's address,
    as do old `sites.simple-host.app/<handle>/<site>/` ones.
    Visitors sign in with Google or an emailed code on the site's own address (a
    sign-in covers that site only); every save from a page needs a signed-in visitor. Agents save with the API key
    (or the connector) through `https://simple-host.app/v1/...`.
  version: 2.0.0
servers:
  - url: https://simple-host.app
...
paths:
...
  /v1/sites/{sitename}/restore:
    post:
      operationId: restoreSite
      summary: Restore a recently deleted site
      description: >-
        Brings a site back from Recently deleted under the same name and address, with every
        version, its saved data, collections and connected address, live at once.
      tags: [Sites]
      security: [{ apiKey: [] }]
      parameters: [ { $ref: '#/components/parameters/sitename' } ]
      responses:
        '200': { description: The restored site, as in GET /v1/sites }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Site quota reached }
        '404': { description: No site with that name in Recently deleted }
        '409': { description: That site is live, not deleted }
...
components:
  securitySchemes:
    apiKey: { type: apiKey, in: header, name: X-API-Key }
    connector:
      type: oauth2
      description: >-
        Per-person OAuth 2.1 (authorization code; PKCE S256 required for every self-registered
        client). An access token sent as `Authorization: Bearer` acts exactly as that person's
        X-API-Key on every route here. Discovery at /.well-known/oauth-authorization-server.
      flows:
        authorizationCode:
          authorizationUrl: https://simple-host.app/oauth/authorize
          tokenUrl: https://simple-host.app/oauth/token
          refreshUrl: https://simple-host.app/oauth/token
          scopes:
            sites: Publish and manage your sites and their data
  parameters:
...
    sitename: { name: sitename, in: path, required: true, schema: { type: string }, example: my-cool-site }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: What went wrong, in plain words. The wording may change; branch on `code`. }
        code:
          type: string
          description: >-
            A stable snake_case name for the refusal, present where one status
            has more than one meaning. site_exists (409: this account already
            has a site of that name), domain_taken (409: the address belongs to
            another site or person), invalid_name (400: not 1-63 lowercase
            letters, digits and hyphens), name_reserved (400: reserved by the
            platform), invalid_domain (400: not a domain that can be connected),
            site_quota_reached (403), append_only (409: public list items cannot
            be edited; the owner may still delete them), custom_domain_required (409: the site needs
            its own address first), not_an_object (409: the saved value is not a
            JSON object), private_visitor_only and private_needs_own_domain
            (403), use_custom_domain and visitor_auth_required (401),
            missing_api_key, wrong_auth_header (sent Authorization instead
            of X-API-Key) and invalid_api_key (unknown, revoked or signed-out
            key) (401), csrf_required (403), recently_deleted (409: the name
            belongs to a site in Recently deleted), site_not_deleted (409:
            restore of a live site), confirm_required (400),
            not_found (404, private collections), export_link_invalid (404:
            the download link expired or is not valid), site_suspended (403:
            the operator has taken the site down; changes and public reads of
            its data are refused until it is restored, `reason` says why),
            account_suspended (403: the operator has suspended this account;
            its keys, connected apps and sign-in are refused, `reason` says
            why). New codes may be added.
          example: domain_taken
  responses:
...
    Unauthorized: { description: "Missing or invalid API key — send the X-API-Key header, not Authorization Bearer. The body's code is missing_api_key, wrong_auth_header or invalid_api_key.", content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...